REF system: kernel skeleton split, frame helpers, lyapunov fix
